% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtach.R
\name{dtach}
\alias{dtach}
\title{detach and unload a package.  usually works the way one might expect it to.}
\usage{
dtach(p)
}
\arguments{
\item{p}{the package name, unquoted like in library()}
}
\value{
whatever base::detach() returns, if p was attached
}
\description{
detach and unload a package.  usually works the way one might expect it to.
}
\details{
If a package is not directly loaded, base::detach will provide helpful
feedback, in the eloquent yet restrained manner of a DEC VAX compiler.
}
\examples{
library(MASS)
dtach(MASS) 

}
