% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{A package for the 2021 edition of Experimental Design and Biostatistics.}
\usage{
.onAttach(lib, pkgname = "ExpDesign2021")
}
\description{
A package for the 2021 edition of Experimental Design and Biostatistics.
}
\examples{

# phelp == 'package help'
phelp(ExpDesign2021) 

# automate fetching assignment data 
fetchAssignments()

# list object sizes (lsos) and pipe the output into head (default: 6 lines)
lsos() \%>\% head 

}
