% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchAssignments.R
\name{fetchAssignments}
\alias{fetchAssignments}
\title{fetch the Assignments spreadsheet for massaging and analysis}
\usage{
fetchAssignments(
 
    url = "https://docs.google.com/spreadsheets/d/1bJw_ad0PLQmLe4RvmRzsVnrOF-9Z9Xth9GTSdwO8B4M"
)
}
\value{
a tibble, via googlesheets4::read_sheet()
}
\description{
fetch the Assignments spreadsheet for massaging and analysis
}
