% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identical.R
\name{identical}
\alias{identical}
\title{A version of identical() that accepts arbitrarily many inputs.
(It turns out that Reduce(identical, ...) doesn't do what you might like.)}
\usage{
identical(...)
}
\arguments{
\item{...}{(list) The things to test for identicality.}
}
\value{
\preformatted{ (boolean) Whether they are all identical.
}
}
\description{
A version of identical() that accepts arbitrarily many inputs.
(It turns out that Reduce(identical, ...) doesn't do what you might like.)
}
\examples{

identical(1:3*1, c(1, 2, 3), seq(3, 9, 3) / 3)           # TRUE
identical(1:3, c(1, 2, 3), seq(1, 3))                    # FALSE
identical(c(1, 2, 3), c(2, 4, 6) / 2, c(3, 6, 9) / 3)    # TRUE 
identical(1:3, c(1, 2, 3), seq_len(3))                   # FALSE 
identical(1:3, seq_len(3), seq_along(1:3))               # TRUE 

}
