% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsos.R
\name{lsos}
\alias{lsos}
\title{list memory-resident objects by footprint}
\usage{
lsos(n = 10, ...)
}
\arguments{
\item{n}{optional integer specifying how many to show (default is 10)}

\item{...}{optional arguments to pass on to .ls.objects()}
}
\value{
a pretty-printed list of objects and their sizes
}
\description{
list memory-resident objects by footprint
}
