% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phelp.R
\name{phelp}
\alias{phelp}
\title{get the list of named documentation files for a package from its namespace}
\usage{
phelp(...)
}
\arguments{
\item{a}{package name(space) as an unquoted string}
}
\value{
an interactive list of documented topics for that package
}
\description{
get the list of named documentation files for a package from its namespace
}
\examples{
help(package="Rscripts") 
phelp(Rscripts) 
phelp(MASS) 

}
