% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.data.frame.R
\name{print.data.frame}
\alias{print.data.frame}
\title{pretty-print data.frames like IRanges or Julia do}
\usage{
\method{print}{data.frame}(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
pretty-print data.frames like IRanges or Julia do
}
