% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqInstall.R
\name{reqInstall}
\alias{reqInstall}
\title{require a package, and if it isn't installed, install it}
\usage{
reqInstall(p)
}
\arguments{
\item{p}{the package to require/install}
}
\value{
whatever require() or install.packages() returns
}
\description{
require a package, and if it isn't installed, install it
}
\examples{
reqInstall("BiocManager") 

}
